Updated browser module
Updated crates' structure
