+++ engine/Cargo.toml
@@ -70,6 +70,7 @@ zeromq = { version = "0.4", default-features = false, features = [
 
 aws-config = { version = "1.1.7", features = ["behavior-version-latest"] }
 aws-sdk-s3 = "1.66.0"
+tempfile = "3.14.0"
 
 [build-dependencies]
 static-files = "0.2"
